Create backend

+++ studentSpring/src/main/java/ru/student/studentSpring/tutorial/dto/course/CoursesParams.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 @Data
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class CoursesParams {
-    private Integer idd;
+    private Long idd;
     private String name;
     private String description;
     private Integer maxCountStudent;
